Support running ormesh as root.
Launching tor as root causes warnings that interfere with control
password hash generation on agent startup. Add support for running tor
as a different user to fix this. Running ormesh as root is necessary
when forwarding privileged ports to remotes.

There's probably more permissions that could be dropped when doing so
but this is a start.